+++ src/modules/common/components/OutboundLink.js
@@ -1,19 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { translate } from 'react-i18next';
+import omit from 'lodash/omit';
+
+import SMIcon from '../../home/components/SMIcon';
+
+const TRANSLATION_PROPS = ['t', 'tReady', 'i18n', 'lng'];
 
 const OutboundLink = translate()(({ t, href, children, ...rest }) => (
   <a
     href={href}
     target="_blank"
     rel="noopener noreferrer"
-    className="outbound-link"
-    {...rest}
+    {...omit(rest, TRANSLATION_PROPS)}
   >
     {children}{' '}
-    <span className="outbound-link__outbound_warning">
-      {t('OUTBOUND_LINK.DESCRIPTION')}
-    </span>
+    <SMIcon icon="outbound-link" title={t('OUTBOUND_LINK.DESCRIPTION')} />
   </a>
 ));